fix(article): fixes alignment for figure image/captions
