+++ src/v2/Apps/Article/Components/Sections/ArticleSectionImageCollection/index.ts
@@ -1,0 +1,1 @@
+export * from "./ArticleSectionImageCollection"
